requirements.txt
django==4.1.7
djangorestframework==3.14.0
spotipy==2.22.1
django-silk==5.0.3
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
django-filter==22.1